+++ src/LookupAnything/release-notes.md
@@ -3,6 +3,8 @@
 # Release notes
 ## 1.19.3
 * Added neutral gifts when looking up an NPC.
+* Debug fields are now listed in alphabetical order.
+* Debug fields now include non-field properties.
 
 ## 1.19.2
 * Improved translations. Thanks to pomepome (Japanese)!